requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
alembic==1.14.0
pydantic==2.10.1
pandas==2.2.3
openpyxl==3.1.5
python-dotenv==1.0.1
pillow==10.4.0
